+++ android/app/src/main/java/app/gridspin/MainActivity.java
@@ -1,0 +1,5 @@
+package app.gridspin;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}